// src/generated/responses.ts
export type GetArtistAppearsOnData = GetArtistData;
export type GetArtistDiscographyOverviewData = GetArtistData;
export type GetArtistDiscoveredOnData = GetArtistData;
export type GetArtistFeaturingData = GetArtistData;
export type GetArtistTopTracksData = GetTracksData;
// ...
export interface ItemClass {
  id: string;
  uri: string;
  profile: UserLocationClass;
  visuals?: ItemVisuals;
  sharingInfo?: PurpleSharingInfo;
  relatedContent?: ItemRelatedContent;
}

export interface UserLocationClass {
  name: string;
}

export interface ItemRelatedContent {
  relatedArtists: RelatedArtists;
}

export interface RelatedArtists {
  totalCount: number;
  items: RelatedArtistsItem[];
}

export interface RelatedArtistsItem {
  id: string;
  uri: string;
  profile: UserLocationClass;
  visuals: ItemVisuals;
}

export interface ItemVisuals {
  avatarImage: AvatarImageElement;
}

export interface AvatarImageElement {
  sources: Icon[];
}

export interface Icon {
  url: string;
  height: number | null;
  width: number | null;
}

export interface PurpleSharingInfo {
  shareUrl: string;
}
// ...
export interface AlbumSharingInfo {
  shareId: string;
  shareUrl: string;
}
// ...
export interface ExternalUrls {
  spotify: string;
}

export enum OwnerType {
  Artist = 'artist',
  User = 'user',
}

export interface AlbumExternalids {
  upc: string;
}

export enum ReleaseDatePrecision {
  Day = 'day',
  Empty = '',
}

export interface AlbumsClass {
  href: string;
  limit: number;
  next: null;
  offset: number;
  previous: null;
  total: number;
  items: TracksTrack[];
}

export interface TracksTrack {
  artists: Owner[];
  disc_number: number;
  track_number: number;
  duration_ms: number;
  explicit: boolean;
  external_urls: ExternalUrls;
  href: string;
  id: string;
  is_local: boolean;
  name: string;
  preview_url: null;
  type: TrackType;
  uri: string;
  external_ids: ItemExternalids;
  album?: ItemElement;
  is_playable?: boolean;
  popularity?: number;
  available_markets?: Market[];
}
// ...
export interface GetArtistAlbumsData {
  artist: GetArtistAlbumsDataArtist;
}

export interface GetArtistAlbumsDataArtist {
  discography: FluffyDiscography;
}

export interface FluffyDiscography {
  albums: PurpleAlbums;
}
// ...
export interface GetArtistData {
  artist: GetArtistAppearsOnDataArtist;
}

export interface GetArtistAppearsOnDataArtist {
  id: string;
  uri: string;
  profile: UserLocationClass;
  relatedContent?: PurpleRelatedContent;
  discography?: TentacledDiscography;
  visuals?: ItemVisuals;
}
// ...
export interface ArtistStats {
  followers: number;
  monthlyListeners: number;
  worldRank: number;
  topCities: TopCities;
}

export interface TopCities {
  items: TopCitiesItem[];
}

export interface TopCitiesItem {
  numberOfListeners: number;
  city: string;
  country: Market;
  region: string;
}

export interface PurpleVisuals {
  gallery: ImagesClass;
  avatarImage: HeaderImageClass;
  headerImage: HeaderImageClass;
}

export interface HeaderImageClass {
  sources: Icon[];
  extractedColors: HeaderImageExtractedColors;
}

export interface HeaderImageExtractedColors {
  colorRaw: Color;
}

export interface GetArtistOverviewData {
  artistUnion: GetArtistOverviewDataArtistUnion;
}

export interface GetArtistOverviewDataArtistUnion {
  __typename: ArtistUnionTypename;
  id: string;
  uri: string;
  sharingInfo: AlbumSharingInfo;
  profile: FluffyProfile;
  visuals: PurpleVisuals;
  discography: IndigoDiscography;
  stats: ArtistStats;
  relatedContent: ArtistUnionRelatedContent;
  goods: PurpleGoods;
}

export enum ArtistUnionTypename {
  Album = 'Album',
  Artist = 'Artist',
  Playlist = 'Playlist',
  Track = 'Track',
  User = 'User',
}

export interface IndigoDiscography {
  latest: LatestElement;
  popularReleasesAlbums: PopularReleasesAlbums;
  singles: CompilationsClass;
  albums: CompilationsClass;
  compilations: CompilationsClass;
  topTracks: PurpleTopTracks;
}

export interface PopularReleasesAlbums {
  totalCount: number;
  items: LatestElement[];
}

export interface PurpleGoods {
  events: Events;
  merch: FluffyMerch;
}

export interface FluffyMerch {
  items: Item2[];
}

export interface Item2 {
  image: BackgroundImageClass;
  name?: string;
  description: string;
  price: string;
  uri: string;
  url: string;
  nameV2?: string;
}

export interface FluffyProfile {
  name: LabelEnum;
  verified: boolean;
  pinnedItem: FluffyPinnedItem;
  biography: BiographyElement;
  externalLinks: ExternalLinks;
  playlistsV2: PlaylistsV2;
}
// ...
export interface ArtistUnionRelatedContent {
  appearsOn: PurpleAppearsOn;
  featuringV2: V2;
  discoveredOnV2: V2;
  relatedArtists: RelatedArtists;
}
// ...
export interface GetArtistRelatedData {
  artist: ItemClass;
}

export interface GetArtistSinglesData {
  artistUnion: GetArtistSinglesDataArtistUnion;
}

export interface GetArtistSinglesDataArtistUnion {
  __typename: ArtistUnionTypename;
  discography: IndecentDiscography;
}

export interface IndecentDiscography {
  singles: Singles;
}
// ...
export interface GetTracksData {
  tracks: TracksTrack[];
}

export interface GetArtistsData {
  artists: GetArtistsDataArtist[];
}

export interface GetArtistsDataArtist {
  external_urls: ExternalUrls;
  followers: Followers;
  genres: any[];
  href: string;
  id: string;
  images: Icon[];
  name: string;
  popularity: number;
  type: OwnerType;
  uri: string;
}

export interface Followers {
  href: null | string;
  total: number | null;
}